+++ tla-bmcmt/src/main/scala/at/forsyte/apalache/tla/bmcmt/rewriter/SymbStateRewriterSnapshot.scala
@@ -9,7 +9,7 @@ import at.forsyte.apalache.tla.lir.TlaEx
 import scala.collection.immutable.SortedSet
 
 class SymbStateRewriterSnapshot(val typeFinderSnapshot: TrivialTypeSnapshot,
-                                val intValueCacheSnapshot: AbstractCacheSnapshot[Arena, Int, ArenaCell],
+                                val intValueCacheSnapshot: AbstractCacheSnapshot[Arena, BigInt, ArenaCell],
                                 val intRangeCacheSnapshot: AbstractCacheSnapshot[Arena, (Int, Int), ArenaCell],
                                 val strValueCacheSnapshot: AbstractCacheSnapshot[Arena, String, ArenaCell],
                                 val recordDomainCache: AbstractCacheSnapshot[Arena, (SortedSet[String], SortedSet[String]), ArenaCell],